pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Vicente Junior", email = "vicente19981@live.com"},
]
dependencies = [
    "djangorestframework>=3.14.0",
    "django>=4.0.4",
]
requires-python = ">=3.10.2"
readme = "README.md"
license = {text = "MIT"}

[project.urls]
Homepage = ""
[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"

[tool.pdm.scripts]
pep8 = "autopep8 --in-place --recursive ./src"
pylint = "pylint ./src"
test = "pytest --ignore __pypackages__"
test_cov = "pdm run test --cov=./src --cov-fail-under=80"
test_cov_html = "pdm run test_cov --cov-report=html:./__coverage"

[tool.pdm.dev-dependencies]
dev = [
    "autopep8>=2.0.2",
    "pylint>=2.17.4",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]
